pyproject.toml
[project]
name = "mp13bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "André Scherer Kunz",email = "andreschererkunzt@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-debug-toolbar (>=5.2.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flake8 (>=7.2.0,<8.0.0)"
]

[tool.poetry]
packages = [
    { include = "MP13Bookstore" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
factory-boy = "^3.3.3"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
